refactor: change location of clients module + minor codestyle things
