+++ src/database/database.module.ts
@@ -3,6 +3,7 @@ import { PrismaService } from './prisma.service';
 import { PairService } from './pair.service';
 import { PairLiquidityInfoHistoryService } from './pair-liquidity-info-history.service';
 import { PairLiquidityInfoHistoryErrorService } from './pair-liquidity-info-history-error.service';
+
 @Module({
   providers: [
     PrismaService,
