+++ src/database/pair.service.ts
@@ -3,6 +3,7 @@ import { Pair, Token } from '@prisma/client';
 import { PrismaService } from './prisma.service';
 
 export type PairWithTokens = { token0: Token; token1: Token } & Pair;
+
 @Injectable()
 export class PairService {
   constructor(private prisma: PrismaService) {}
