+++ src/tasks/tasks.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
 import { PairLiquidityInfoHistoryImporterService } from './pair-liquidity-info-history-importer.service';
 import { DatabaseModule } from '../database/database.module';
-import { MdwClientModule } from '../clients/mdw/mdw-client.module';
+import { ClientsModule } from '../clients/clients.module';
 import { ScheduleModule } from '@nestjs/schedule';
 
 @Module({
-  imports: [MdwClientModule, DatabaseModule, ScheduleModule.forRoot()],
+  imports: [ClientsModule, DatabaseModule, ScheduleModule.forRoot()],
   providers: [PairLiquidityInfoHistoryImporterService],
 })
 export class TasksModule {}
